# apps/aurora-core/build/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "aurora-core",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "/api"
		}
	],
	"components": {
...
		"securitySchemes": {
			"local": {
				"type": "apiKey",
				"name": "connect.sid",
				"in": "cookie"
			},
			"integration": {
				"type": "apiKey",
				"name": "x-api-key",
				"in": "header"
			}
		},
		"schemas": {
...
			"Partial_Record_SocketioNamespaces.string__": {
				"properties": {
					"/": {
						"type": "string"
					},
					"/audio": {
						"type": "string"
					},
					"/screen": {
						"type": "string"
					},
					"/lights": {
						"type": "string"
					},
					"/public": {
						"type": "string"
					},
					"/backoffice": {
						"type": "string"
					},
					"/backoffice-beat": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"LightsControllerResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"socketIds": {
						"$ref": "#/components/schemas/Partial_Record_SocketioNamespaces.string__",
						"description": "ID of the socket connection if present.\nRequired to send events specifically and only to this entity.\nMapping from namespace to ID, as a websocket has a different ID\nfor each namespace its in."
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LightsControllerCreateParams": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FooterSize": {
				"enum": [
					"full",
					"minimal",
					"hidden"
				],
				"type": "string"
			},
			"BasePoster": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"due": {
						"type": "string",
						"format": "date-time"
					},
					"timeout": {
						"type": "number",
						"format": "double"
					},
					"footer": {
						"$ref": "#/components/schemas/FooterSize"
					},
					"borrelMode": {
						"type": "boolean",
						"description": "Whether this poster should only be shown when in BorrelMode"
					},
					"color": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"label",
					"timeout",
					"footer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PosterType.AGENDA": {
				"enum": [
					"agenda"
				],
				"type": "string"
			},
			"PosterType.INFIMA": {
				"enum": [
					"infima"
				],
				"type": "string"
			},
			"PosterType.LOGO": {
				"enum": [
					"logo"
				],
				"type": "string"
			},
			"PosterType.TRAINS": {
				"enum": [
					"train"
				],
				"type": "string"
			},
			"PosterType.BORREL_LOGO": {
				"enum": [
					"borrel-logo"
				],
				"type": "string"
			},
			"PosterType.BORREL_PRICE_LIST": {
				"enum": [
					"borrel-price-list"
				],
				"type": "string"
			},
			"PosterType.BORREL_WALL_OF_SHAME": {
				"enum": [
					"borrel-wall-of-shame"
				],
				"type": "string"
			},
			"PosterType.OLYMPICS": {
				"enum": [
					"olympics"
				],
				"type": "string"
			},
			"PosterType.UNKNOWN": {
				"enum": [
					"unknown"
				],
				"type": "string"
			},
			"LocalPosterType": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/PosterType.AGENDA"
					},
					{
						"$ref": "#/components/schemas/PosterType.INFIMA"
					},
					{
						"$ref": "#/components/schemas/PosterType.LOGO"
					},
					{
						"$ref": "#/components/schemas/PosterType.TRAINS"
					},
					{
						"$ref": "#/components/schemas/PosterType.BORREL_LOGO"
					},
					{
						"$ref": "#/components/schemas/PosterType.BORREL_PRICE_LIST"
					},
					{
						"$ref": "#/components/schemas/PosterType.BORREL_WALL_OF_SHAME"
					},
					{
						"$ref": "#/components/schemas/PosterType.OLYMPICS"
					},
					{
						"$ref": "#/components/schemas/PosterType.UNKNOWN"
					}
				]
			},
			"LocalPoster": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BasePoster"
					},
					{
						"properties": {
							"type": {
								"$ref": "#/components/schemas/LocalPosterType"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					}
				]
			},
			"PosterType.IMAGE": {
				"enum": [
					"img"
				],
				"type": "string"
			},
			"PosterType.VIDEO": {
				"enum": [
					"video"
				],
				"type": "string"
			},
			"PosterType.EXTERNAL": {
				"enum": [
					"extern"
				],
				"type": "string"
			},
			"MediaPoster": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BasePoster"
					},
					{
						"properties": {
							"source": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"type": {
								"anyOf": [
									{
										"$ref": "#/components/schemas/PosterType.IMAGE"
									},
									{
										"$ref": "#/components/schemas/PosterType.VIDEO"
									},
									{
										"$ref": "#/components/schemas/PosterType.EXTERNAL"
									}
								]
							}
						},
						"required": [
							"source",
							"type"
						],
						"type": "object"
					}
				]
			},
			"PosterType.PHOTO": {
				"enum": [
					"photo"
				],
				"type": "string"
			},
			"PhotoPoster": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BasePoster"
					},
					{
						"properties": {
							"albums": {
								"items": {
									"type": "number",
									"format": "double"
								},
								"type": "array"
							},
							"type": {
								"$ref": "#/components/schemas/PosterType.PHOTO"
							}
						},
						"required": [
							"albums",
							"type"
						],
						"type": "object"
					}
				]
			},
			"PosterType.ERROR": {
				"enum": [
					"error"
				],
				"type": "string"
			},
			"ErrorPoster": {
				"allOf": [
					{
						"$ref": "#/components/schemas/BasePoster"
					},
					{
						"properties": {
							"message": {
								"type": "string"
							},
							"type": {
								"$ref": "#/components/schemas/PosterType.ERROR"
							}
						},
						"required": [
							"message",
							"type"
						],
						"type": "object"
					}
				]
			},
			"Poster": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/LocalPoster"
					},
					{
						"$ref": "#/components/schemas/MediaPoster"
					},
					{
						"$ref": "#/components/schemas/PhotoPoster"
					},
					{
						"$ref": "#/components/schemas/ErrorPoster"
					}
				]
			},
			"PosterResponse": {
				"properties": {
					"posters": {
						"items": {
							"$ref": "#/components/schemas/Poster"
						},
						"type": "array"
					},
					"borrelMode": {
						"type": "boolean"
					}
				},
				"required": [
					"posters",
					"borrelMode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BorrelModeResponse": {
				"properties": {
					"enabled": {
						"type": "boolean"
					},
					"present": {
						"type": "boolean"
					}
				},
				"required": [
					"enabled",
					"present"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BorrelModeParams": {
				"properties": {
					"enabled": {
						"type": "boolean"
					}
				},
				"required": [
					"enabled"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrainResponse": {
				"properties": {
					"direction": {
						"type": "string"
					},
					"plannedDateTime": {
						"type": "string"
					},
					"delay": {
						"type": "number",
						"format": "double"
					},
					"trainType": {
						"type": "string"
					},
					"operator": {
						"type": "string"
					},
					"cancelled": {
						"type": "boolean"
					},
					"routeStations": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"messages": {
						"items": {
							"properties": {
								"message": {
									"type": "string"
								},
								"style": {
									"type": "string"
								}
							},
							"required": [
								"message",
								"style"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"direction",
					"plannedDateTime",
					"delay",
					"trainType",
					"operator",
					"cancelled",
					"routeStations",
					"messages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PhotoResponse": {
				"properties": {
					"label": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GEWISPhotoAlbumParams": {
				"properties": {
					"albumIds": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"albumIds"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MedalTableRecord": {
				"properties": {
					"countryName": {
						"type": "string"
					},
					"flagUrl": {
						"type": "string"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"gold": {
						"type": "number",
						"format": "double"
					},
					"silver": {
						"type": "number",
						"format": "double"
					},
					"bronze": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"countryName",
					"flagUrl",
					"gold",
					"silver",
					"bronze",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OlympicMedal": {
				"properties": {
					"participantId": {
						"type": "number",
						"format": "double"
					},
					"eventId": {
						"type": "number",
						"format": "double"
					},
					"participantName": {
						"type": "string"
					},
					"sportName": {
						"type": "string"
					},
					"eventName": {
						"type": "string"
					},
					"medal": {
						"type": "string",
						"enum": [
							"gold",
							"silver",
							"bronze"
						]
					}
				},
				"required": [
					"participantId",
					"eventId",
					"participantName",
					"sportName",
					"eventName",
					"medal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CountryMedalResponse": {
				"properties": {
					"countryName": {
						"type": "string"
					},
					"flagUrl": {
						"type": "string"
					},
					"rank": {
						"type": "number",
						"format": "double"
					},
					"gold": {
						"type": "number",
						"format": "double"
					},
					"silver": {
						"type": "number",
						"format": "double"
					},
					"bronze": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"medals": {
						"items": {
							"$ref": "#/components/schemas/OlympicMedal"
						},
						"type": "array"
					}
				},
				"required": [
					"countryName",
					"flagUrl",
					"gold",
					"silver",
					"bronze",
					"total",
					"medals"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		}
	},
	"paths": {
...
		"/lights/controller": {
			"get": {
				"operationId": "GetLightsControllers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LightsControllerResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateLightsController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsControllerResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LightsControllerCreateParams"
							}
						}
					}
				}
			}
		},
		"/lights/controller/{id}": {
			"get": {
				"operationId": "GetSingleLightsController",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LightsControllerResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lights"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/handler/screen/poster/carousel": {
			"get": {
				"operationId": "GetPosters",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PosterResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "alwaysReturnBorrelPosters",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/handler/screen/poster/carousel/force-update": {
			"post": {
				"operationId": "ForceUpdatePosters",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico"
						]
					}
				],
				"parameters": []
			}
		},
		"/handler/screen/poster/carousel/borrel-mode": {
			"get": {
				"operationId": "GetPosterBorrelMode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BorrelModeResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "SetPosterBorrelMode",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BorrelModeParams"
							}
						}
					}
				}
			}
		},
		"/handler/screen/poster/carousel/train-departures": {
			"get": {
				"operationId": "GetTrains",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TrainResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
		"/handler/screen/poster/carousel/photo": {
			"post": {
				"operationId": "GetPhoto",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PhotoResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GEWISPhotoAlbumParams"
							}
						}
					}
				}
			}
		},
		"/handler/screen/poster/carousel/olympics/medal-table": {
			"get": {
				"operationId": "GetOlympicsMedalTable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MedalTableRecord"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
		"/handler/screen/poster/carousel/olympics/country-medals": {
			"get": {
				"operationId": "GetDutchOlympicMedals",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CountryMedalResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Handlers"
				],
				"security": [
					{
						"local": [
							"admin",
							"board",
							"avico",
							"bac",
							"key-holder",
							"audio-subscriber",
							"screen-subscriber",
							"lights-subscriber",
							"integration-user"
						]
					}
				],
				"parameters": []
			}
		},
...
	}
}